+++ ljepotaservis/ljepotaservis.Web/ClientApp/src/components/ownerSide/index.js
@@ -1,15 +1,28 @@
 import React from "react";
+import { connect } from "react-redux";
+import { Route, Switch, Redirect } from "react-router";
 import Dashboard from "./Dashboard";
 import Employees from "./Employees";
-import { Route, Switch, Redirect } from "react-router";
+import RoleNavbar from "../navbar";
 
-export default function OwnerSide(props) {
+const OwnerSide = props => {
   const { path } = props.match;
   return (
-    <Switch>
-      <Route exact path={path} component={Dashboard} />
-      <Route path={path + "/employees"} component={Employees} />
-      <Redirect to={path} />
-    </Switch>
+    <div>
+      <RoleNavbar role={props.role} />
+      <Switch>
+        <Route exact path={path} component={Dashboard} />
+        <Route path={path + "/employees"} component={Employees} />
+        <Redirect to={path} />
+      </Switch>
+    </div>
   );
-}
+};
+
+const mapStateToProps = state => ({
+  role: state.authentication.hasOwnProperty("user")
+    ? state.authentication.user.role
+    : "Guest"
+});
+
+export default connect(mapStateToProps)(OwnerSide);
