+++ ljepotaservis/ljepotaservis.Web/ClientApp/src/components/superAdminSide/index.js
@@ -1,15 +1,28 @@
 import React from "react";
+import { connect } from "react-redux";
+import { Route, Switch, Redirect } from "react-router";
 import Dashboard from "./Dashboard";
 import Stores from "./Stores";
-import { Route, Switch, Redirect } from "react-router";
+import RoleNavbar from "../navbar";
 
-export default function SuperAdminSide(props) {
+const SuperAdminSide = props => {
   const { path } = props.match;
   return (
-    <Switch>
-      <Route exact path={path} component={Dashboard} />
-      <Route path={path + "/stores"} component={Stores} />
-      <Redirect to={path} />
-    </Switch>
+    <div>
+      <RoleNavbar role={props.role} />
+      <Switch>
+        <Route exact path={path} component={Dashboard} />
+        <Route path={path + "/stores"} component={Stores} />
+        <Redirect to={path} />
+      </Switch>{" "}
+    </div>
   );
-}
+};
+
+const mapStateToProps = state => ({
+  role: state.authentication.hasOwnProperty("user")
+    ? state.authentication.user.role
+    : "Guest"
+});
+
+export default connect(mapStateToProps)(SuperAdminSide);
